requirements.txt
gunicorn
fastapi
uvicorn[standard]
azure-search-documents
azure-core[aio]
aiohttp
openai
pydantic
python-dotenv
